requirements.txt
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.9
joblib==1.4.2
pillow==10.4.0
praw==7.7.1
prawcore==2.4.0
proto-plus==1.24.0
protobuf==5.28.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.1.4
python-dotenv==1.0.1
regex==2024.9.11
requests==2.32.3
rsa==4.9
soupsieve==2.6
tqdm==4.66.5
update-checker==0.18.0
uritemplate==4.1.1
urllib3==2.2.3
websocket-client==1.8.0
